requirements.txt
# Core Python packages
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Machine Learning and AI
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0
scikit-learn>=1.0.0
tensorflow>=2.8.0
keras>=2.8.0

# Deep Learning and Neural Networks
transformers>=4.20.0
datasets>=2.0.0
accelerate>=0.15.0
optuna>=3.0.0

# Reinforcement Learning
gym>=0.21.0
stable-baselines3>=1.5.0
ray[rllib]>=2.0.0

# Computer Vision
opencv-python>=4.5.0
Pillow>=8.3.0
imageio>=2.9.0

# Scientific Computing
sympy>=1.10.0
networkx>=2.6.0
igraph>=0.9.0

# Data Processing
h5py>=3.6.0
pyyaml>=6.0
toml>=0.10.0
configparser>=5.2.0

# Visualization
plotly>=5.0.0
bokeh>=2.4.0
dash>=2.0.0

# Utilities
tqdm>=4.62.0
joblib>=1.1.0
psutil>=5.8.0
pygame>=2.1.0

# Testing
pytest>=6.2.0
pytest-cov>=3.0.0
pytest-mock>=3.6.0

# Development
black>=22.0.0
flake8>=4.0.0
mypy>=0.950
pre-commit>=2.17.0

# Documentation
sphinx>=4.5.0
sphinx-rtd-theme>=1.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.4.0

# Audio Processing (for voice commands)
librosa>=0.9.0
soundfile>=0.10.0

# 3D Graphics and Physics
pygame>=2.1.0
pyglet>=1.5.0
pymunk>=6.4.0

# Networking and Communication
websockets>=10.0
aiohttp>=3.8.0
requests>=2.28.0

# Database
sqlalchemy>=1.4.0
sqlite3

# Security
cryptography>=3.4.0
bcrypt>=3.2.0

# System Integration
psutil>=5.8.0
pywin32>=303; sys_platform == 'win32'
pyobjc-framework-Cocoa>=8.0; sys_platform == 'darwin'

# Additional AI/ML packages
xgboost>=1.5.0
lightgbm>=3.3.0
catboost>=1.0.0
shap>=0.40.0
lime>=0.2.0

# Time series and forecasting
prophet>=1.1.0
statsmodels>=0.13.0

# Optimization
cvxpy>=1.2.0
pulp>=2.6.0

# Parallel processing
multiprocessing-logging>=0.3.0
concurrent-futures>=3.1.0

# Logging and monitoring
structlog>=21.5.0
prometheus-client>=0.13.0

# Configuration management
hydra-core>=1.2.0
omegaconf>=2.2.0

# Type checking
typing-extensions>=4.0.0 